Cleanup pkg/db.
Removed the db integration test.
