+++ Makefile
@@ -25,7 +25,6 @@ build_integration_test:
 	@go build -o ./bin/test_domainowner_pca_policlog_interaction  ./tests/integration/domainowner_pca_policlog_interaction
 	@go build -o ./bin/test_mapserver  ./tests/integration/mapserver
 	@go build -o ./bin/test_smt  ./tests/integration/smt
-	@go build -o ./bin/test_db  ./tests/integration/db
 	@go build -o ./bin/test_grpc  ./tests/integration/grpc_test
 
 drop_cacheTable:
